refactor: 타입 컴포넌트 이동
