+++ FE/src/pages/Onboarding/Type/Type.jsx
