+++ FE/src/pages/Onboarding/Type/Type.module.scss
